Adjustments to RawInput Code
