+++ Sparky-core/Sparky-core.vcxproj.filters
@@ -420,7 +420,7 @@
     <ClCompile Include="src\sp\platform\windows\Win32MemoryManager.cpp">
       <Filter>sp\platform\win32</Filter>
     </ClCompile>
-    <ClCompile Include="src\sp\platform\windows\Win32RawInput.cpp" />
+    <ClCompile Include="src\sp\platform\windows\Win32Input.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\sp\app\Application.h">
