+++ Sparky-core/src/sp/platform/windows/Win32Input.cpp
@@ -8,24 +8,122 @@
 #include "sp/app/Window.h"
 #include "sp/events/Events.h"
 
+#define RAW_BUFFER_SIZE 50
+
 namespace sp {
 
 	using namespace events;
 
 	extern HWND hWnd;
+	extern HRAWINPUT rawInputHandle;
+
+	static byte rawBuffer[RAW_BUFFER_SIZE];
+
+	void InputManager::PlatformInit()
+	{
+		RAWINPUTDEVICE rid[2];
+
+		rid[0].usUsagePage = 1;
+		rid[0].usUsage = 6;
+		rid[0].dwFlags = 0;
+		rid[0].hwndTarget = NULL;
+
+		rid[1].usUsagePage = 1;
+		rid[1].usUsage = 2;
+		rid[1].dwFlags = 0;
+		rid[1].hwndTarget = NULL;
+		if (!RegisterRawInputDevices(rid, 2, sizeof(rid[0])))
+		{
+			SP_ERROR("Failed registering raw input devices!");
+		}
+	}
 
 	void InputManager::PlatformUpdate()
 	{
-		// Mouse Events
-		POINT mouse;
-		GetCursorPos(&mouse);
-		ScreenToClient(hWnd, &mouse);
+		m_MouseDelta = m_MouseDeltaCurrent;
+		m_MouseDeltaCurrent = 0;
+	}
 
-		maths::vec2 mousePos = maths::vec2((float)mouse.x, (float)mouse.y);
-		if (mousePos != m_MousePosition)
+	void InputManager::PlatformUpdateMessage()
+	{
+		uint bufferSize = RAW_BUFFER_SIZE;
+		GetRawInputData(rawInputHandle, RID_INPUT, (LPVOID)rawBuffer, &bufferSize, sizeof(RAWINPUTHEADER));
+
+		RAWINPUT* raw = (RAWINPUT*)rawBuffer;
+
+		if (raw->header.dwType == RIM_TYPEMOUSE)
 		{
-			m_EventCallback(MouseMovedEvent(mousePos.x, mousePos.y, m_MouseButtons[SP_MOUSE_LEFT]));
-			m_MousePosition = mousePos;
+			if ((bool)(m_MouseClicked[SP_MOUSE_1] = raw->data.mouse.ulButtons & RI_MOUSE_BUTTON_1_DOWN)) m_MouseButtons[SP_MOUSE_1] = true;
+			else if ((bool)(raw->data.mouse.ulButtons & RI_MOUSE_BUTTON_1_UP))
+			{
+				m_MouseButtons[SP_MOUSE_1] = false;
+				m_EventCallback(MouseReleasedEvent(SP_MOUSE_1, m_MousePosition.x, m_MousePosition.y));
+			}
+
+			if ((bool)(m_MouseClicked[SP_MOUSE_2] = raw->data.mouse.ulButtons & RI_MOUSE_BUTTON_2_DOWN)) m_MouseButtons[SP_MOUSE_2] = true;
+			else if ((bool)(raw->data.mouse.ulButtons & RI_MOUSE_BUTTON_2_UP))
+			{
+				m_MouseButtons[SP_MOUSE_2] = false;
+				m_EventCallback(MouseReleasedEvent(SP_MOUSE_2, m_MousePosition.x, m_MousePosition.y));
+			}
+
+			if ((bool)(m_MouseClicked[SP_MOUSE_3] = raw->data.mouse.ulButtons & RI_MOUSE_BUTTON_3_DOWN)) m_MouseButtons[SP_MOUSE_3] = true;
+			else if ((bool)(raw->data.mouse.ulButtons & RI_MOUSE_BUTTON_3_UP))
+			{
+				m_MouseButtons[SP_MOUSE_3] = false;
+				m_EventCallback(MouseReleasedEvent(SP_MOUSE_3, m_MousePosition.x, m_MousePosition.y));
+			}
+
+			if ((bool)(m_MouseClicked[SP_MOUSE_4] = raw->data.mouse.ulButtons & RI_MOUSE_BUTTON_4_DOWN)) m_MouseButtons[SP_MOUSE_4] = true;
+			else if ((bool)(raw->data.mouse.ulButtons & RI_MOUSE_BUTTON_4_UP))
+			{
+				m_MouseButtons[SP_MOUSE_4] = false;
+				m_EventCallback(MouseReleasedEvent(SP_MOUSE_4, m_MousePosition.x, m_MousePosition.y));
+			}
+
+			if ((bool)(m_MouseClicked[SP_MOUSE_5] = raw->data.mouse.ulButtons & RI_MOUSE_BUTTON_5_DOWN)) m_MouseButtons[SP_MOUSE_5] = true;
+			else if ((bool)(raw->data.mouse.ulButtons & RI_MOUSE_BUTTON_5_UP))
+			{
+				m_MouseButtons[SP_MOUSE_5] = false;
+				m_EventCallback(MouseReleasedEvent(SP_MOUSE_5, m_MousePosition.x, m_MousePosition.y));
+			}
+
+			for (int i = SP_MOUSE_1; i <= SP_MOUSE_5; i++)
+			{
+				if (m_MouseClicked[i]) m_EventCallback(MousePressedEvent(i, m_MousePosition.x, m_MousePosition.y));
+			}
+
+			if (raw->data.mouse.lLastX != 0 || raw->data.mouse.lLastY != 0)
+			{
+				m_MousePosition.x += raw->data.mouse.lLastX;
+				m_MousePosition.y += raw->data.mouse.lLastY;
+
+				m_MouseDeltaCurrent.x += raw->data.mouse.lLastX;
+				m_MouseDeltaCurrent.y += raw->data.mouse.lLastY;
+
+				m_EventCallback(MouseMovedEvent(m_MousePosition.x, m_MousePosition.y, m_MouseButtons[SP_MOUSE_1]));
+			}
+
+			if (raw->data.mouse.ulButtons & RI_MOUSE_WHEEL)
+			{
+				m_EventCallback(MousePressedEvent(raw->data.mouse.usButtonData == 120 ? SP_MWHEEL_UP : SP_MWHEEL_DOWN, m_MousePosition.x, m_MousePosition.y));
+			}
+		}
+
+		if (raw->header.dwType == RIM_TYPEKEYBOARD)
+		{
+			if (raw->data.keyboard.Flags & RI_KEY_BREAK)
+			{
+
+				m_KeyState[raw->data.keyboard.VKey] = false;
+				m_EventCallback(KeyReleasedEvent(raw->data.keyboard.VKey));
+			}
+			else
+			{
+				bool repeat = m_KeyState[raw->data.keyboard.VKey];
+				m_KeyState[raw->data.keyboard.VKey] = true;
+				m_EventCallback(KeyPressedEvent(raw->data.keyboard.VKey, repeat, 0));
+			}
 		}
 	}
 
@@ -40,94 +138,17 @@ namespace sp {
 	{
 		if (cursor == SP_NO_CURSOR)
 		{
+			RECT wrc;
+			GetWindowRect(hWnd, &wrc);
+			ClipCursor(&wrc);
 			SetCursor(SP_NO_CURSOR);
 			while (ShowCursor(false) >= 0);
 		}
 		else
 		{
 			SetCursor(LoadCursor(NULL, IDC_ARROW));
+			ClipCursor(NULL);
 			ShowCursor(true);
 		}
 	}
-
-	void KeyCallback(InputManager* inputManager, int32 flags, int32 key, uint message)
-	{
-		bool pressed = message == WM_KEYDOWN || message == WM_SYSKEYDOWN;
-		inputManager->m_KeyState[key] = pressed;
-
-		bool repeat = (flags >> 30) & 1;
-
-		int32 modifier = 0;
-		switch (key)
-		{
-		case SP_KEY_CONTROL:
-			modifier = SP_MODIFIER_LEFT_CONTROL;
-			break;
-		case SP_KEY_ALT:
-			modifier = SP_MODIFIER_LEFT_ALT;
-			break;
-		case SP_KEY_SHIFT:
-			modifier = SP_MODIFIER_LEFT_SHIFT;
-			break;
-		}
-		if (pressed)
-			inputManager->m_KeyModifiers |= modifier;
-		else
-			inputManager->m_KeyModifiers &= ~(modifier);
-
-		if (pressed)
-			inputManager->m_EventCallback(KeyPressedEvent(key, repeat, inputManager->m_KeyModifiers));
-		else
-			inputManager->m_EventCallback(KeyReleasedEvent(key));
-	}
-
-	void MouseButtonCallback(InputManager* inputManager, int32 button, int32 x, int32 y)
-	{
-		bool down = false;
-		switch (button)
-		{
-		case WM_LBUTTONDOWN:
-			SetCapture(hWnd);
-			button = SP_MOUSE_LEFT;
-			down = true;
-			break;
-		case WM_LBUTTONUP:
-			ReleaseCapture();
-			button = SP_MOUSE_LEFT;
-			down = false;
-			break;
-		case WM_RBUTTONDOWN:
-			SetCapture(hWnd);
-			button = SP_MOUSE_RIGHT;
-			down = true;
-			break;
-		case WM_RBUTTONUP:
-			ReleaseCapture();
-			button = SP_MOUSE_RIGHT;
-			down = false;
-			break;
-		case WM_MBUTTONDOWN:
-			SetCapture(hWnd);
-			button = SP_MOUSE_MIDDLE;
-			down = true;
-			break;
-		case WM_MBUTTONUP:
-			ReleaseCapture();
-			button = SP_MOUSE_MIDDLE;
-			down = false;
-			break;
-		}
-		inputManager->m_MouseButtons[button] = down;
-		inputManager->m_MousePosition.x = (float)x;
-		inputManager->m_MousePosition.y = (float)y;
-
-		SP_ASSERT(inputManager->m_EventCallback);
-
-		if (down)
-			inputManager->m_EventCallback(MousePressedEvent(button, (float)x, (float)y));
-		else
-			inputManager->m_EventCallback(MouseReleasedEvent(button, (float)x, (float)y));
-	}
-
-
 }
